chore(codegen): split some utils in 'java' and 'scala' variants
Should also fix some errors on master due to parallel merges
